% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR3.R
\name{cost}
\alias{cost}
\title{Cost function described in equation (2) of SR3 paper}
\usage{
cost(A, b, x, lambda = 1)
}
\arguments{
\item{A}{m x d real valued matrix representing
linear data generating mechanism}

\item{b}{m dimensional vector of observations}

\item{x}{d dimensional vector of unknown signal}

\item{lambda}{Scalar to control strength of regularization}
}
\value{
Loss of least squares problem regularized by
R with strength lambda
}
\description{
Cost function described in equation (2) of SR3 paper
}
