% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR3.R
\name{G_kappa}
\alias{G_kappa}
\title{Function to compute matrix G_k}
\usage{
G_kappa(A, HkInv, kappa)
}
\arguments{
\item{A}{m x d real valued matrix representing
linear data generating mechanism}

\item{HkInv}{Inverse of matrix Hk}

\item{kappa}{Relaxation parameter to control gap
between x and w}
}
\value{
Matrix with shape (m + d) x m
}
\description{
Function to compute matrix G_k
}
