% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR3.R
\name{Hk}
\alias{Hk}
\title{Function to compute d x d invertible matrix H_k}
\usage{
Hk(A, kappa)
}
\arguments{
\item{A}{m x d real valued matrix representing
linear data generating mechanism}

\item{kappa}{Relaxation parameter to control gap
between x and w}
}
\value{
d x d matrix
}
\description{
Function to compute d x d invertible matrix H_k
}
