% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l012Rprox.R
\name{norm_vec}
\alias{norm_vec}
\title{Return either the value of the penalty or the solution of the prox problem (if ifprox)}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{Vector on which prox operator is run}

\item{alpha}{weight}

\item{mode}{string, '0' -> p = 0, etc}

\item{ifprox}{flag}
}
\description{
Return either the value of the penalty or the solution of the prox problem (if ifprox)
}
