% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR3.R
\name{R}
\alias{R}
\title{Compute L1 regularization}
\usage{
R(x, lambda = 1)
}
\arguments{
\item{x}{A vector of weights}

\item{lambda}{Parameter to control the amount of regularization}
}
\value{
The L1 norm of \code{x} weighed uniformly by \code{lambda}.
}
\description{
Compute L1 regularization
}
