% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr3.R
\name{reg_prox}
\alias{reg_prox}
\title{Function to return prox operator and regularization function}
\usage{
reg_prox(p)
}
\arguments{
\item{p}{Parser list with input values}
}
\description{
Function to return prox operator and regularization function
}
