% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR3.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Proximal operator for L1 regularized least squares}
\usage{
soft_threshold(x, eta, lambda = 1)
}
\arguments{
\item{x}{A vector of weights}

\item{eta}{Learning rate}

\item{lambda}{Parameter to control the amount of regularization}
}
\value{
Matrix with shape (m + d) x d
}
\description{
Proximal operator for L1 regularized least squares
}
