% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr3.R
\name{reg_prox}
\alias{reg_prox}
\title{Prox operators}
\usage{
reg_prox(p, alpha)
}
\arguments{
\item{p}{The argument to \code{sr3_parse_input} which is a list that
contains the mode variable for regularization}
}
\value{
A list of \code{R} and \code{Rprox}
}
\description{
Function to return prox operator and regularization function
}
